+++ lib/data/constraints.dart
@@ -1,1 +1,2 @@
 const double kDefaultHeaderHeight = 40;
+const double kReorderHandleWidth = 40;