+++ lib/vit_table.dart
@@ -1,6 +1,7 @@
 library vit_table;
 
 export 'package:vit_table/data/models/vit_table_column.dart';
+export 'package:vit_table/data/models/vit_table_reorder_mode.dart';
 export 'package:vit_table/data/models/vit_table_row.dart';
 export 'package:vit_table/ui/components/organisms/vit_table.dart';
 export 'package:vit_table/ui/theme/vit_table_style.dart';